requirements.txt
fastapi
uvicorn
jinja2
python-dotenv
python-multipart
sqlalchemy
groq
requests

